Move assertion helpers to `pytest_helpers.py`
