+++ array_api_tests/test_statistical_functions.py
@@ -1,5 +1,5 @@
 import math
-from typing import Optional, Tuple
+from typing import Optional
 
 from hypothesis import assume, given
 from hypothesis import strategies as st
@@ -11,50 +11,12 @@ from . import hypothesis_helpers as hh
 from . import pytest_helpers as ph
 from . import shape_helpers as sh
 from . import xps
-from .typing import DataType, Scalar, ScalarType, Shape
+from .typing import DataType
 
 
 def kwarg_dtypes(dtype: DataType) -> st.SearchStrategy[Optional[DataType]]:
     dtypes = [d2 for d1, d2 in dh.promotion_table if d1 == dtype]
     return st.none() | st.sampled_from(dtypes)
-
-
-def assert_keepdimable_shape(
-    func_name: str,
-    out_shape: Shape,
-    in_shape: Shape,
-    axes: Tuple[int, ...],
-    keepdims: bool,
-    /,
-    **kw,
-):
-    if keepdims:
-        shape = tuple(1 if axis in axes else side for axis, side in enumerate(in_shape))
-    else:
-        shape = tuple(side for axis, side in enumerate(in_shape) if axis not in axes)
-    ph.assert_shape(func_name, out_shape, shape, **kw)
-
-
-def assert_equals(
-    func_name: str,
-    type_: ScalarType,
-    idx: Shape,
-    out: Scalar,
-    expected: Scalar,
-    /,
-    **kw,
-):
-    out_repr = "out" if idx == () else f"out[{idx}]"
-    f_func = f"{func_name}({ph.fmt_kw(kw)})"
-    if type_ is bool or type_ is int:
-        msg = f"{out_repr}={out}, should be {expected} [{f_func}]"
-        assert out == expected, msg
-    elif math.isnan(expected):
-        msg = f"{out_repr}={out}, should be {expected} [{f_func}]"
-        assert math.isnan(out), msg
-    else:
-        msg = f"{out_repr}={out}, should be roughly {expected} [{f_func}]"
-        assert math.isclose(out, expected, rel_tol=0.25, abs_tol=1), msg
 
 
 @given(
@@ -72,7 +34,7 @@ def test_max(x, data):
 
     ph.assert_dtype("max", x.dtype, out.dtype)
     _axes = sh.normalise_axis(kw.get("axis", None), x.ndim)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "max", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     scalar_type = dh.get_scalar_type(out.dtype)
@@ -83,7 +45,7 @@ def test_max(x, data):
             s = scalar_type(x[idx])
             elements.append(s)
         expected = max(elements)
-        assert_equals("max", scalar_type, out_idx, max_, expected)
+        ph.assert_scalar_equals("max", scalar_type, out_idx, max_, expected)
 
 
 @given(
@@ -101,7 +63,7 @@ def test_mean(x, data):
 
     ph.assert_dtype("mean", x.dtype, out.dtype)
     _axes = sh.normalise_axis(kw.get("axis", None), x.ndim)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "mean", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     for indices, out_idx in zip(sh.axes_ndindex(x.shape, _axes), sh.ndindex(out.shape)):
@@ -112,7 +74,7 @@ def test_mean(x, data):
             s = float(x[idx])
             elements.append(s)
         expected = sum(elements) / len(elements)
-        assert_equals("mean", float, out_idx, mean, expected)
+        ph.assert_scalar_equals("mean", float, out_idx, mean, expected)
 
 
 @given(
@@ -130,7 +92,7 @@ def test_min(x, data):
 
     ph.assert_dtype("min", x.dtype, out.dtype)
     _axes = sh.normalise_axis(kw.get("axis", None), x.ndim)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "min", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     scalar_type = dh.get_scalar_type(out.dtype)
@@ -141,7 +103,7 @@ def test_min(x, data):
             s = scalar_type(x[idx])
             elements.append(s)
         expected = min(elements)
-        assert_equals("min", scalar_type, out_idx, min_, expected)
+        ph.assert_scalar_equals("min", scalar_type, out_idx, min_, expected)
 
 
 @given(
@@ -189,7 +151,7 @@ def test_prod(x, data):
         _dtype = dtype
     ph.assert_dtype("prod", x.dtype, out.dtype, _dtype)
     _axes = sh.normalise_axis(kw.get("axis", None), x.ndim)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "prod", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     scalar_type = dh.get_scalar_type(out.dtype)
@@ -204,7 +166,7 @@ def test_prod(x, data):
         if dh.is_int_dtype(out.dtype):
             m, M = dh.dtype_ranges[out.dtype]
             assume(m <= expected <= M)
-        assert_equals("prod", scalar_type, out_idx, prod, expected)
+        ph.assert_scalar_equals("prod", scalar_type, out_idx, prod, expected)
 
 
 @given(
@@ -236,7 +198,7 @@ def test_std(x, data):
     out = xp.std(x, **kw)
 
     ph.assert_dtype("std", x.dtype, out.dtype)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "std", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     # We can't easily test the result(s) as standard deviation methods vary a lot
@@ -287,7 +249,7 @@ def test_sum(x, data):
         _dtype = dtype
     ph.assert_dtype("sum", x.dtype, out.dtype, _dtype)
     _axes = sh.normalise_axis(kw.get("axis", None), x.ndim)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "sum", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     scalar_type = dh.get_scalar_type(out.dtype)
@@ -302,7 +264,7 @@ def test_sum(x, data):
         if dh.is_int_dtype(out.dtype):
             m, M = dh.dtype_ranges[out.dtype]
             assume(m <= expected <= M)
-        assert_equals("sum", scalar_type, out_idx, sum_, expected)
+        ph.assert_scalar_equals("sum", scalar_type, out_idx, sum_, expected)
 
 
 @given(
@@ -334,7 +296,7 @@ def test_var(x, data):
     out = xp.var(x, **kw)
 
     ph.assert_dtype("var", x.dtype, out.dtype)
-    assert_keepdimable_shape(
+    ph.assert_keepdimable_shape(
         "var", out.shape, x.shape, _axes, kw.get("keepdims", False), **kw
     )
     # We can't easily test the result(s) as variance methods vary a lot
